Major performance boost to dump-kcore.py (#1) Previously, when transferring data from /proc/kcore to the output file, a simple unbuffered read and write was performed. This obviously resulted in very poor performance, because in most scenarios the largest virtual memory range being copied was larger than the current available memory, which results in swapped out pages which impact performance drastically.
Now, instead of a simple read and write, the `copy_file_range` syscall
is utilized to perform the copy within the kernel,
with all of its optimizations.

The end result is a tremendous performance improvement.
So much so, that the performance is now better than AVML
(which is written in rust and uses /proc/kcore as well).

For reference:
on my test machine, with 4GB of RAM,
dumping with AVML took 23-28 seconds,
and dumping with dump-kcore.py took 17-19 seconds!